Setup CMake to fetch boost in order to make it available
This could be done with offline availability, by simply adding boost in dependencies directory
But it seems the build will be done on a POSIX machine with network and this streamlines the process.

Choice of Boost is to use Asio for tcp handling instead of sys/socket.h because it will result in better maintainability.
The tradeoff here is a bigger binary (which may or may not be acceptable).

Binary size is being mitigated by using boost in header-only mode.